+++ src/Fable.AST/Plugins.fs
@@ -35,6 +35,7 @@ type CompilerOptions =
         TrimRootModule: bool
         Verbosity: Verbosity
         FileExtension: string
+        TriggeredByDependency: bool
     }
 
 type PluginHelper =