+++ src/Fable.Transforms/FSharp2Fable.Util.fs
@@ -1750,9 +1750,13 @@ module Util =
             | None when info.IsInterface ->
                 callInstanceMember com r typ callInfo ent memb |> Some
             | None ->
-                com.WarnOnlyOnce("Fable only supports a subset of standard .NET API, please check https://fable.io/docs/dotnet/compatibility.html. For external libraries, check whether they are Fable-compatible in the package docs.")
-                sprintf "%s.%s is not supported by Fable" info.DeclaringEntityFullName info.CompiledName
-                |> addErrorAndReturnNull com ctx.InlinePath r |> Some
+                let msg =
+                    if info.DeclaringEntityFullName.StartsWith("Fable.Core.") then
+                        $"{info.DeclaringEntityFullName}.{info.CompiledName} is not supported, try updating fable tool"
+                    else
+                        com.WarnOnlyOnce("Fable only supports a subset of standard .NET API, please check https://fable.io/docs/dotnet/compatibility.html. For external libraries, check whether they are Fable-compatible in the package docs.")
+                        $"{info.DeclaringEntityFullName}.{info.CompiledName} is not supported by Fable"
+                msg |> addErrorAndReturnNull com ctx.InlinePath r |> Some
         | _ -> None
 
     let addWatchDependencyFromMember (com: Compiler) (memb: FSharpMemberOrFunctionOrValue) =